+++ PBL3_CoffeeHome/BLL/MenuItemBLL.cs
@@ -7,50 +7,30 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.VisualBasic.ApplicationServices;
+using System.Collections;
 
 namespace PBL3_CoffeeHome.BLL
 {
     public class MenuItemBLL
     {
         private readonly MenuItemDAL _menuItemDAL;
+        private readonly OrderDAL _orderDAL;
+        private readonly BaristaQueueBLL _baristaQueueBLL;
         private readonly InventoryTransactionDAL _inventoryTransactionBLL;
         private readonly MenuItemIngredientBLL _menuItemIngredientBLL;
 
         public MenuItemBLL()
         {
             _menuItemDAL = new MenuItemDAL();
+            _orderDAL = new OrderDAL();
+            _baristaQueueBLL = new BaristaQueueBLL();
             _menuItemIngredientBLL = new MenuItemIngredientBLL();
             _inventoryTransactionBLL = new InventoryTransactionDAL();
         }
-        // Lấy danh sách tất cả đơn hàng
-        public List<Order> GetOrders()
-        {
-            return _menuItemDAL.GetAllOrders();
-
-        }
-
-        // Lấy danh sách đơn hàng của ngày hôm nay
-        public List<Order> GetOrdersAssignedToday(string status)
-        {
-            return _menuItemDAL.GetOrdersAssignedToday(status);
-        }
-
-        // Lấy danh sách đơn hàng đã hoàn thành trong ngày được chọn
-        public List<Order> GetOrdersCompletedOnDate(string status, DateTime selectedDate)
-        {
-            return _menuItemDAL.GetOrdersCompletedOnDate(status, selectedDate);
-        }
-
         // Lấy danh sách đơn hàng có trạng thái "Incompleted"
         public List<Order> GetOrdersWithStatus(string status)
         {
-            return _menuItemDAL.GetOrdersWithStatus(status);
-        }
-
-        // Lấy chi tiết các món trong đơn hàng
-        public List<OrderItem> GetOrderDetails(string orderId)
-        {
-            return _menuItemDAL.GetOrderItemsByOrderId(orderId);
+            return _orderDAL.GetOrdersByBaristaQueueStatus(status);
         }
         // Lấy thông tin món theo tên
         public MenuItems GetMenuItemByName(string name)
@@ -59,11 +39,9 @@ namespace PBL3_CoffeeHome.BLL
         }
 
         // Cập nhật trạng thái đơn hàng
-        public void CompleteOrder(string orderId, string userId)
+        public void CompleteOrder(string orderId, string queueID,string userId)
         {
-            var order = _menuItemDAL.GetOrderById(orderId);
-
-            var orderItems = _menuItemDAL.GetOrderItemsByOrderId(orderId);
+            var orderItems = _orderDAL.GetOrderItemsByOrderId(orderId);
             foreach (var orderItem in orderItems)
             {
                 var ingredients = _menuItemIngredientBLL.GetMenuItemIngredient(orderItem.MenuItemID);
@@ -74,7 +52,7 @@ namespace PBL3_CoffeeHome.BLL
                         $"Xuất tự động cho đơn hàng");
                 }
             }
-            _menuItemDAL.UpdateOrderStatus(orderId, "Completed");
+            _baristaQueueBLL.UpdateQueueStatus(queueID, "Completed");
         }
     }
 }
